% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaao_stats.R
\name{calc_iaao_stats}
\alias{calc_iaao_stats}
\title{Evaluates the accuracy of assessments.}
\usage{
calc_iaao_stats(ratios)
}
\arguments{
\item{ratios}{A dataframe which has been pre-processed by \code{\link{reformat_data}}}
}
\value{
Various statistics and facts on assessments by year
}
\description{
Evaluates the accuracy of assessments.
}
\examples{
calc_iaao_stats(ratios)
}
