% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{diagnostic_plots}
\alias{diagnostic_plots}
\title{Produces plots and tables helpful for data diagnostics}
\usage{
diagnostic_plots(stats, ratios, min_reporting_yr, max_reporting_yr)
}
\arguments{
\item{stats}{the output of calc_iaao_stats}

\item{ratios}{the output of reformat_data}

\item{min_reporting_yr}{the minimum year to plot}

\item{max_reporting_yr}{the maximum year to ploy}
}
\value{
a list of plots (cod, prd, prb, binned_scatter)
}
\description{
Produces plots and tables helpful for data diagnostics
}
