% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example assessment and sale data}
\format{
A dataset with 4 columns and 346,164 rows.
\describe{
\item{PIN}{A unique identifier for each property}
\item{ASSESSED_VALUE}{The assessed value for the property}
\item{SALE_YEAR}{The year of the property's sale}
\item{SALE_PRICE}{The sale amount}
}
}
\source{
\url{https://datacatalog.cookcountyil.gov/}
}
\usage{
example_data
}
\description{
A dataset containing assessment and sale data from Cook County, Illinois. This data was modified from its original form.
}
\keyword{datasets}
