% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo.R
\name{monte_carlo_graphs}
\alias{monte_carlo_graphs}
\title{Conducts a monte carlo simulation, which approximates the impacts of random factors in regressivity statistics}
\usage{
monte_carlo_graphs(ratios)
}
\arguments{
\item{ratios}{a dataframe with assessment, sales, and time data which has been processed by reformat_data}
}
\value{
a list of six graphs for COD, PRB, PRD, paglin72, cheng74, and IAAO78 (see README for example)
}
\description{
Conducts a monte carlo simulation, which approximates the impacts of random factors in regressivity statistics
}
