% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo.R
\name{paglin_cheng_IAAO_coefs}
\alias{paglin_cheng_IAAO_coefs}
\title{Conducts a monte carlo simulation, which approximates the impacts of random factors in regressivity statistics}
\usage{
paglin_cheng_IAAO_coefs(ratios)
}
\arguments{
\item{ratios}{a dataframe with assessment, sales, and time data which has been processed by reformat_data}

\item{return_model_objs}{boolean, if true will return a list of model objects}

\item{produce_table}{boolean, if true will return a summary table produced using stargazer}
}
\value{
a list of model objects or a summary data.frame of results
}
\description{
Conducts a monte carlo simulation, which approximates the impacts of random factors in regressivity statistics
}
