% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_over_under.R
\name{pct_over_under}
\alias{pct_over_under}
\title{Produces a key visualization of assessment regressivity, the percent over-under plot}
\usage{
pct_over_under(ratios, min_reporting_yr, max_reporting_yr, jurisdiction_name)
}
\arguments{
\item{ratios}{the output of reformat_data}

\item{min_reporting_yr}{the minimum year to plot}

\item{max_reporting_yr}{the maximum year to ploy}

\item{jurisdiction_name}{the name of the jurisdiction analyzed}
}
\value{
a list (caption, plot)
}
\description{
Produces a key visualization of assessment regressivity, the percent over-under plot
}
