% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plots}
\alias{plots}
\title{Produces various plots}
\usage{
plots(stats, ratios, min_reporting_yr, max_reporting_yr)
}
\arguments{
\item{stats}{the output of iaao_stats}

\item{ratios}{the output of reformat_data}

\item{min_reporting_yr}{the minimum year to plot}

\item{max_reporting_yr}{the maximum year to ploy}
}
\value{
a list of plots (cod, prd, prb, binned_scatter)
}
\description{
Produces various plots
}
