% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_tests.R
\name{regression_tests}
\alias{regression_tests}
\title{Offers various tests for regressivity. See https://erhla.github.io/cmfproperty/Methods for more info.}
\usage{
regression_tests(ratios, return_model_objs = FALSE, produce_table = FALSE)
}
\arguments{
\item{ratios}{a dataframe with assessment, sales, and time data which has been processed by reformat_data}

\item{return_model_objs}{boolean, if true will return a list of model objects}

\item{produce_table}{boolean, if true will return a summary table produced using stargazer}
}
\value{
a summary data.frame of results or, if return_model_objs is True a list of model objects
}
\description{
Offers various tests for regressivity. See https://erhla.github.io/cmfproperty/Methods for more info.
}
