% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_tests.R
\name{vertical_equity_metrics}
\alias{vertical_equity_metrics}
\title{Offers various tests for regressivity.}
\usage{
vertical_equity_metrics(property_data, return_model_objs = FALSE)
}
\arguments{
\item{property_data}{a dataframe with assessment, sales, and time data which has been processed by reformat_data}

\item{return_model_objs}{boolean, if true will return a list of model objects
#' @return a list of model objects or a summary data.frame of results}
}
\description{
Offers various tests for regressivity.
}
