% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaao_graphs.R
\name{iaao_graphs}
\alias{iaao_graphs}
\title{Produces graphs of IAAO standards over time}
\usage{
iaao_graphs(
  stats,
  ratios,
  min_reporting_yr,
  max_reporting_yr,
  jurisdiction_name
)
}
\arguments{
\item{stats}{the output of iaao_stats}

\item{ratios}{the output of reformat_data}

\item{min_reporting_yr}{the minimum year to plot}

\item{max_reporting_yr}{the maximum year to ploy}

\item{jurisdiction_name}{the name of the jurisdiction analyzed}
}
\value{
a list of plots (cod, prd, prb, binned_scatter)
}
\description{
Produces graphs of IAAO standards over time
}
