% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprocess_data.R
\name{reformat_data}
\alias{reformat_data}
\title{Reformats data for analysis by changing column names and filtering out non-arm's length sales.}
\usage{
reformat_data(
  df,
  sale_col,
  assessment_col,
  sale_year_col,
  filter_data,
  market_value_col = NULL,
  tax_year_col = NULL
)
}
\arguments{
\item{df}{a dataframe with assessment, sales, and time data}

\item{sale_col}{the name of the column with sales data}

\item{assessment_col}{the name of the column with assessment data}

\item{sale_year_col}{the name of the column with the year of sale}

\item{filter_data}{if True, keeps only arm's length sales. if False, keeps all data.}

\item{market_value_col}{optional, the name of the column with market value data}

\item{tax_year_col}{optional, the name of the column with tax year data. The default is to set to the sale year.}
}
\value{
a dataframe with appropriate column names and arm's length markings
}
\description{
Reformats data for analysis by changing column names and filtering out non-arm's length sales.
}
