% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_tests.R
\name{regression_tests}
\alias{regression_tests}
\title{Offers various tests for regressivity.}
\usage{
regression_tests(property_data, return_model_objs = FALSE)
}
\arguments{
\item{property_data}{a dataframe with assessment, sales, and time data which has been processed by reformat_data}

\item{return_model_objs}{boolean, if true will return a list of model objects}
}
\value{
a list of model objects or a summary data.frame of results
}
\description{
Offers various tests for regressivity.
}
